requirements.txt
discord.py
addict
requests